Combine Utils into TOWCommon in Utilities namespace
